requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
idna==3.10
python-decouple==3.8
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
